Homogenize comments

+++ include/cart.h
@@ -16,10 +16,43 @@ typedef struct {
 } rom_header;
 
 // Cart save/load functions
+
+/**
+ * @brief Loads a cart into the context.
+ *
+ * @param cart the filename of the cart to load.
+ * @return true if the cart is loaded successfully.
+ * @return false if the cart failed to load.
+ */
 bool cart_load(char *cart);
+
+/**
+ * @brief Loads the game state from a .sav file into cart ram.
+ *
+ */
 void cart_load_ram();
+
+/**
+ * @brief Saves the game state from cart ram into a .sav file.
+ *
+ */
 void cart_save_ram();
 
 // Cart RAM functions
+
+/**
+ * @brief Reads a byte from the cart ram.
+ *
+ * @param addr the address to read from.
+ * @return uint8_t the value at the address.
+ */
 void write_cart_ram(uint32_t addr, uint8_t val);
-uint16_t read_cart_ram(uint32_t addr);
+
+
+/**
+ * @brief Writes a byte to the cart ram.
+ *
+ * @param addr the address to write to.
+ * @param val the value to write.
+ */
+uint8_t read_cart_ram(uint32_t addr);
